UPG : config viewmodel improvements
* refactoring finished - all things related to options are placed to the config folder
* separate file for base-types for options
* 'me' instead of 'o' in templates
* template for Tab (with 'message' property)
* Property 'message' for tab
* CSS: column-span mixin
* improved sorting of Blocks (config)
