+++ data/interfaces/default/css/config.less
@@ -104,3 +104,7 @@
 	   -moz-column-rule: @rule;
 	        column-rule: @rule;
 }
+.columnSpan(@what) {
+	-webkit-column-span: @what;
+	        column-span: @what;
+}
