+++ data/interfaces/default/css/style.less
@@ -120,6 +120,14 @@ small, .small
         .columnRule( 1px solid @border-color );
         .columnCount(2);
         .columnGap(4em);
+
+        .message
+        {
+            .columnSpan(all);
+            display:block;
+            background-color:transparent;
+            margin-bottom:1em;
+        }
     }
 
     .block
